refactor: 로또 번호 중복을 검사하는 로직을 lotto 클래스로 이동
